Support for apache calcite merged with main (#10)
* Add working TPCH Q10 query for Apache Calcite.

* Add working query for TPCH Q11 for Apache Calcite. Had to change dynamic parameter type from double to BigDecimal to avoid remote driver exception.

* Working TPCH Q8 query for Apache Calcite.

* Fix TPCH Q1 for Apache Calcite so it works with dynamic parameters.

* Add TPCH Q3 for Apache Calcite with dynamic parameters.

* Use dynamic parameters for TPCH Q4 for Apache Calcite.

* Re-add TPCH Q5 for Apache Calcite.

* Use dynamic parameters for TPCH Q6 for Apache Calcite.

* Add working TPCH Q7 for Apache Calcite with dynamic parameters.

* Add working TPCH Q9 for Apache Calcite with dynamic parameters.

* Add working TPCH Q12 for Apache Calcite with dynamic parameters.

* Add working TPCH Q13 for Apache Calcite with dynamic parameters.

* Use dynamic paramters for TPCH Q14 for Apache Calcite.

* Implement working TPCH Q17 for Apache Calcite.

* Update TPCH Q20 for Apache Calcite to use dynamic parameters.

* Implement TPCH Q21 for Apache Calcite.

* Implement TPCH Q22 for Apache Calcite.

+++ src/main/java/com/oltpbenchmark/benchmarks/tpch/procedures/Q11.java
@@ -22,6 +22,7 @@ import com.oltpbenchmark.benchmarks.tpch.TPCHConstants;
 import com.oltpbenchmark.benchmarks.tpch.TPCHUtil;
 import com.oltpbenchmark.util.RandomGenerator;
 
+import java.math.BigDecimal;
 import java.sql.Connection;
 import java.sql.PreparedStatement;
 import java.sql.SQLException;
@@ -63,11 +64,11 @@ public class Q11 extends GenericQuery {
         String nation = TPCHUtil.choice(TPCHConstants.N_NAME, rand);
 
         // FRACTION is chosen as 0.0001 / SF
-        double fraction = 0.0001 / scaleFactor;
+        BigDecimal fraction = BigDecimal.valueOf(0.0001 / scaleFactor);
 
         PreparedStatement stmt = this.getPreparedStatement(conn, query_stmt);
         stmt.setString(1, nation);
-        stmt.setDouble(2, fraction);
+        stmt.setBigDecimal(2, fraction);
         stmt.setString(3, nation);
         return stmt;
     }
